+++ docs/news.rst
@@ -11,6 +11,8 @@
 Psycopg 3.2.13 (unreleased)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Show the host name in the error message in case of name resolution error
+  (:ticket:`#1205`).
 - Fix `Cursor.copy()` and `AsyncCursor.copy()` to hold the connection lock
   for the entire operation, preventing concurrent access issues (:ticket:`#1210`).
 
@@ -23,8 +25,6 @@ Psycopg 3.2.12
 
 - Allow copy to pass different types per column, as long as the database can
   convert them. Regression introduced in 3.2.11 (:ticket:`#1192`).
-- Show the host name in the error message in case of name resolution error
-  (:ticket:`#1205`).
 
 
 Psycopg 3.2.11